==> carl_lehman_la/notebooks/01_data_cleaning_test.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
from sklearn.model_selection import cross_val_score, KFold, train_test_split
from sklearn.preprocessing import StandardScaler
from sklearn.linear_model import Ridge, Lasso, ElasticNet, LinearRegression, RidgeCV, LassoCV, ElasticNetCV
from sklearn.metrics import mean_squared_error
%matplotlib inline

df = pd.read_csv('../data/test (1).csv' , index_col = 'Id')

# Exploring the dataset

df.info()

df.describe().T

df.head()

# Data Cleaning

df.loc[df['Garage Yr Blt'] == 2207.0, 'Garage Yr Blt'] = np.mean(df['Garage Yr Blt'])

columns_na = {'Lot Frontage': 0.0,
              'Alley': 'No Alley',
              'Mas Vnr Type': 'None',
              'Mas Vnr Area': 0.0,
              'Bsmt Qual': 'No Basement',
              'Bsmt Cond': 'No Basement',
              'Bsmt Exposure': 'No Basement',
              'BsmtFin Type 1': 'No Basement',
              'BsmtFin SF 1': 0.0,
              'BsmtFin Type 2': 'No Basement',
              'BsmtFin SF 2': 0.0,
              'Bsmt Unf SF': 0.0,
              'Total Bsmt SF': 0.0,
              'Bsmt Full Bath': 0.0,
              'Bsmt Half Bath': 0.0,
              'Fireplace Qu': 'No Fireplace',
              'Garage Type': 'No Garage',
              'Garage Yr Blt': np.mean(df['Garage Yr Blt']),
              'Garage Finish': 'No Garage',
              'Garage Cars': 0.0,
              'Garage Area': 0.0,
              'Garage Qual': 'No Garage',
              'Garage Cond': 'No Garage',
              'Pool QC': 'No Pool',
              'Fence': 'No Fence',
              'Misc Feature': 'None'}


df.fillna(value = columns_na, inplace=True)

df.isnull().sum().sum()

# Creating Dummy Variables

 Based on the scatter plots made in the exploratory data analysis section, I picked out the numerical,
 categorical variables and put them in a list in order to turn them into objects so I can make 
 dummy columns out of them

columns_to_object = ['MS SubClass', 'Overall Qual', 'Overall Cond', 'Bsmt Full Bath', 'Bsmt Half Bath',
                  'Full Bath', 'Half Bath', 'Bedroom AbvGr', 'Kitchen AbvGr',
                  'Fireplaces', 'Garage Cars', 'Mo Sold', 'Yr Sold']

for col in columns_to_object:
    df[col] = df[col].astype(object)
df.info()

df_object = pd.get_dummies(df[list(df.select_dtypes(include="object"))], 
                           prefix = list(df.select_dtypes(include="object").columns), 
                           columns = df.select_dtypes(include="object").columns)
df = pd.concat([df, df_object], axis = 1, join = 'outer')

# Dropping columns that were turned into dummies

df.drop(['MS SubClass', 'Bsmt Full Bath', 'Bsmt Half Bath',
         'Full Bath', 'Half Bath', 'Bedroom AbvGr', 'Kitchen AbvGr', 'TotRms AbvGrd',
         'Fireplaces', 'Mo Sold', 'Garage Cars', 'Yr Sold', 'BsmtFin SF 1',
         'BsmtFin SF 2', 'Misc Val'], axis=1, inplace=True)

# Save the cleaned data

df.to_csv('../data/clean_data_test.csv')
