==> carl_lehman_la/notebooks/02_Pre-Processing.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
from sklearn.model_selection import cross_val_score, KFold, train_test_split
from sklearn.preprocessing import StandardScaler
from sklearn.linear_model import Ridge, Lasso, ElasticNet, LinearRegression, RidgeCV, LassoCV, ElasticNetCV
from sklearn.metrics import mean_squared_error
from sklearn.feature_selection import VarianceThreshold
import pickle
import csv
np.random.seed(42)
%matplotlib inline

# Load in cleaned train data and cleaned test data

df = pd.read_csv('../data/clean_data.csv', index_col = 'Id') 

test = pd.read_csv('../data/clean_data_test.csv' ,index_col = 'Id')

df.head()

test.head()

# Creating a function that spits out a set difference

def Diff(li1, li2):
    return (list(set(li1) - set(li2)))

Creating columns in df found in test data but not found in train data

for x in Diff(list(test.columns), list(df.columns)):
    df[x] = 0

Creating columns in test dataframe found in train data but not in test data

for x in Diff(list(df.columns), list(test.columns)):
    if x != 'SalePrice':
        test[x] = 0

test.shape

df.shape

# Setting my X and y

X = df.drop('SalePrice', 1)
y = df['SalePrice']

# Train test split

X_train, X_test, y_train, y_test = train_test_split(X,y, random_state= 42)

# Setting and fitting a variance threshold

The variance threshold gets rid of columns with low variance, helping us get rid of features

# Variance Threshold
threshold = VarianceThreshold(0.05)
X_train_var = threshold.fit_transform(X_train.select_dtypes(exclude = 'object'))
X_train.select_dtypes(exclude='object').columns[threshold.get_support()]

# Variance Threshold
X_test_var = threshold.transform(X_test.select_dtypes(exclude='object'))

# Scaling my data

ss = StandardScaler()

X_train_sc = ss.fit_transform(X_train_var)
X_test_sc = ss.transform(X_test_var)

# Exporting my data and scaler for future use

with open('../data/X_train_sc.csv', 'w+') as f:
    csv_writer = csv.writer(f)
    csv_writer.writerows(X_train_sc)

with open('../data/X_test_sc.csv', 'w+') as f:
    csv_writer = csv.writer(f)
    csv_writer.writerows(X_test_sc)

with open('../data/X_train.csv', 'w+') as f:
    csv_writer = csv.writer(f)
    csv_writer.writerows(X_train_var)

with open('../data/X_test.csv', 'w+') as f:
    csv_writer = csv.writer(f)
    csv_writer.writerows(X_test_var)

y_train.to_csv('../data/y_train.csv', index=False, header=False)

y_test.to_csv('../data/y_test.csv', index=False, header=False)

with open('../assets/scaler.pkl', 'wb+') as f:
    pickle.dump(ss, f)

# Exporting my columns for future use

with open('../assets/columns.pkl', 'wb+') as f:
    pickle.dump(list(X_train.select_dtypes(exclude = 'object').columns[threshold.get_support()]), f)

test.to_csv('../data/kaggle.csv')

df.to_csv('../data/df_clean_final.csv')
