==> carl_lehman_la/notebooks/03_Modeling-ridge.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
from sklearn.model_selection import cross_val_score, KFold, train_test_split
from sklearn.preprocessing import StandardScaler
from sklearn.linear_model import Ridge, Lasso, ElasticNet, LinearRegression, RidgeCV, LassoCV, ElasticNetCV
from sklearn.metrics import mean_squared_error
import pickle
import csv
from sklearn.pipeline import Pipeline
from sklearn.grid_search import GridSearchCV
from sklearn.feature_selection import SelectKBest, VarianceThreshold, f_regression
np.random.seed(42)
%matplotlib inline

# Importing my already scaled data

X_train_sc = pd.read_csv('../data/X_train_sc.csv')
X_test_sc = pd.read_csv('../data/X_test_sc.csv')
y_train = pd.read_csv('../data/y_train.csv')
y_test = pd.read_csv('../data/y_test.csv')

# Instantiaing  and fitting my model

ridge = RidgeCV(alphas = np.logspace(0,5,200))

ridge.fit(X_train_sc, y_train)

# Scoring my model

ridge.score(X_train_sc,y_train)

ridge.score(X_test_sc, y_test)

# Calculating RMSE for my model

y_hat_train = ridge.predict(X_train_sc)
y_hat_test = ridge.predict(X_test_sc)

mean_squared_error(y_train, y_hat_train) ** .5

mean_squared_error(y_test, y_hat_test) ** .5

# Creating a dataframe of coefficients and their weights

with open('../assets/columns.pkl', 'rb') as f:
    columns = pickle.load(f)

coef = pd.DataFrame(ridge.coef_.T, index = columns, columns=['weight'])
coef.sort_values(by='weight', ascending=False).head(20)

coef = pd.DataFrame(ridge.coef_.T, index = columns, columns=['weight'])
coef.sort_values(by='weight', ascending=True).head(20)

# Saving my model for future use

with open('../assets/ridge.pkl', 'wb+') as f:
    pickle.dump(ridge, f)
