==> carl_lehman_la/notebooks/04_Submission_lasso.ipynb <==
import pandas as pd
# import numpy as np
# import matplotlib.pyplot as plt
# import seaborn as sns
# from sklearn.model_selection import cross_val_score, KFold, train_test_split
# from sklearn.preprocessing import StandardScaler
# from sklearn.metrics import mean_squared_error
import pickle
# import csv
# np.random.seed(42)
# %matplotlib inline

# Loading in my model, scaler, columns, and data

with open('../assets/scaler.pkl', 'rb') as f:
    ss = pickle.load(f)
    
with open('../assets/lasso.pkl', 'rb') as f:
    lasso = pickle.load(f)
    
with open('../assets/columns.pkl', 'rb') as f:
    columns = pickle.load(f)

kaggle = pd.read_csv('../data/kaggle.csv',  index_col='Id')
df = pd.read_csv('../data/df_clean_final.csv', index_col = 'Id')

kaggle.shape

df.shape

# Putting my test data in the right order

kaggle = kaggle[columns]

len(kaggle[columns].columns)

kaggle_sc = ss.transform(kaggle.select_dtypes(exclude = 'object'))#.drop(['BsmtFin SF 1',
                                                                        #'Overall Qual',
                                                                        #'Overall Cond'], axis=1))

# Creating my predictions

SalePrice = lasso.predict(X = kaggle_sc)

kaggle['SalePrice'] = SalePrice

# Putting my predictions into the correct format for submission

data = {'SalePrice': kaggle['SalePrice']}

submit = pd.DataFrame(data)

submit.sort_values(by = 'Id', axis = 0).head(10)

lasso.intercept_

submit.to_csv('../data/Submit.csv')
